RavenDB-20268 Adjust namespaces of termproviders.
