+++ src/Corax/IndexSearcher/MultiTermMatches/IndexSearcher.MultiTermMatch.Simple.cs
@@ -2,6 +2,7 @@
 using System.Text.RegularExpressions;
 using Corax.Mappings;
 using Corax.Queries;
+using Corax.Queries.TermProviders;
 using Voron;
 using Voron.Data.CompactTrees;
 using Voron.Data.Lookups;
