+++ src/Corax/Queries/Meta/RangeMarker.cs
@@ -1,0 +1,8 @@
+namespace Corax.Queries.Meta;
+
+public static class Range
+{
+    public interface Marker { }
+    public struct Exclusive : Marker { }
+    public struct Inclusive : Marker { }
+}
