+++ src/Corax/Queries/TermProviders/TermProvider.EndsWith.cs
@@ -4,9 +4,8 @@ using Corax.Mappings;
 using Voron;
 using Voron.Data.CompactTrees;
 using Voron.Data.Lookups;
-using CompactTreeForwardIterator = Voron.Data.CompactTrees.CompactTree.Iterator<Voron.Data.Lookups.Lookup<Voron.Data.CompactTrees.CompactTree.CompactKeyLookup>.ForwardIterator>;
 
-namespace Corax.Queries
+namespace Corax.Queries.TermProviders
 {
     public struct EndsWithTermProvider<TLookupIterator> : ITermProvider
         where TLookupIterator : struct, ILookupIterator
