+++ src/Corax/Queries/TermProviders/TermProvider.Exists.cs
@@ -1,14 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using Corax.Mappings;
-using Sparrow;
-using Voron;
 using Voron.Data.CompactTrees;
 using Voron.Data.Lookups;
-using CompactTreeForwardIterator = Voron.Data.CompactTrees.CompactTree.Iterator<Voron.Data.Lookups.Lookup<Voron.Data.CompactTrees.CompactTree.CompactKeyLookup>.ForwardIterator>;
 
-namespace Corax.Queries
+namespace Corax.Queries.TermProviders
 {
     public struct ExistsTermProvider<TLookupIterator> : ITermProvider
         where TLookupIterator : struct, ILookupIterator
