+++ src/Corax/Queries/TermProviders/TermProvider.NumericRange.cs
@@ -1,0 +1,106 @@
+﻿using System.Collections.Generic;
+using System.Diagnostics;
+using Corax.Mappings;
+using Corax.Queries.Meta;
+using Voron.Data.Lookups;
+
+
+namespace Corax.Queries.TermProviders
+{
+    [DebuggerDisplay("{DebugView,nq}")]
+    public struct TermNumericRangeProvider<TLookupIterator, TLow, THigh, TVal> : ITermProvider
+        where TLookupIterator : struct, ILookupIterator
+        where TLow : struct, Range.Marker
+        where THigh  : struct, Range.Marker
+        where TVal : struct, ILookupKey
+    {
+        private readonly IndexSearcher _searcher;
+        private readonly FieldMetadata _field;
+        private readonly TVal _low, _high;
+        private TLookupIterator _iterator;
+        private bool _first;
+
+        public TermNumericRangeProvider(IndexSearcher searcher, Lookup<TVal> set, FieldMetadata field, TVal low, TVal high)
+        {
+            _searcher = searcher;
+            _field = field;
+            _iterator = set.Iterate<TLookupIterator>();
+            _low = low;
+            _high = high;
+            _first = true;
+        }
+
+        public void Reset()
+        {
+            _first = true;
+        }
+        
+        public bool Next(out TermMatch term)
+        {
+            bool wasFirst = false;
+            if (_first)
+            {
+                _iterator.Seek(_iterator.IsForward ? _low : _high);
+                _first = false;
+                wasFirst = true;
+            }
+
+            bool hasNext = _iterator.MoveNext(out TVal key, out var termId);
+            if (hasNext == false)
+                goto Empty;
+
+
+            if (wasFirst)
+            {
+                if (_iterator.IsForward && typeof(TLow) == typeof(Range.Exclusive) && key.IsEqual(_low))
+                    return Next(out term);
+
+                if (_iterator.IsForward == false && typeof(THigh) == typeof(Range.Exclusive) &&
+                    _high.CompareTo(key) <= 0) // in case of (x, y) we've to skip item up to y-double.Eps
+                    return Next(out term);
+
+                if (_iterator.IsForward == false && typeof(THigh) == typeof(Range.Inclusive) && _high.CompareTo(key) < 0)
+                    return Next(out term);
+            }
+            
+            //In case of going forward we've to compare with the highest (right element). In case of backward we've to compare to the lowest element.
+            if (_iterator.IsForward)
+            {
+                var cmp = _high.CompareTo(key);
+                if (typeof(THigh) == typeof(Range.Exclusive) && cmp <= 0 ||
+                    typeof(THigh) == typeof(Range.Inclusive) && cmp < 0)
+                {
+                    goto Empty;
+                }
+            }
+            else
+            {
+                var cmp = _low.CompareTo(key);
+                if (typeof(TLow) == typeof(Range.Exclusive) && cmp >= 0 ||
+                    typeof(TLow) == typeof(Range.Inclusive) && cmp > 0)
+                    goto Empty;
+            }
+            
+            // Ratio will be always 1 (sizeof(T)/sizeof(T))
+            term = _searcher.TermQuery(_field, termId, 1);
+            return true;
+
+            Empty:
+            term = TermMatch.CreateEmpty(_searcher, _searcher.Allocator);
+            return false;
+        }
+
+        public QueryInspectionNode Inspect()
+        {
+            return new QueryInspectionNode($"{GetType().Name}",
+                            parameters: new Dictionary<string, string>()
+                            {
+                                { "Field", _field.ToString() },
+                                { "Low", _low.ToString()},
+                                { "High", _high.ToString()}
+                            });
+        }
+
+        public string DebugView => Inspect().ToString();
+    }
+}
